adds branching dialogue paths based on facelift game score -- temporary 5s play time
to have this work, must add goalFrog to Scene1, attach getBones script to frog2_manybones and goalFrog, and adjust translation vector in Scoring to the distance to the goalFrog.

Then add an empty object called Facelift to Scene, add component Facelift.cs and Scoring.cs

then you should be able to run.